+++ test/test_ul4.py
@@ -4771,6 +4771,17 @@ def test_method_mimeformat_datetime(T):
 
 
 @pytest.mark.ul4
+def test_method_timestamp_datetime(T):
+	t2 = datetime.datetime(1970, 1, 1)
+
+	assert '-3600.0' == T("<?print data.timestamp()?>").renders(data=t2)
+	assert '-3600.0' == T("<?code m = data.timestamp?><?print m()?>").renders(data=t2)
+
+	with raises(argumentmismatchmessage):
+		T("<?print @(2000-02-29T12:34:56.987000).timestamp(42)?>").renders()
+
+
+@pytest.mark.ul4
 def test_method_date_date(T):
 	assert '2000-02-29' == T('<?print @(2000-02-29).date()?>').renders()