+++ docs/NEWS.rst
@@ -17,6 +17,8 @@ Changes in HEAD (released 2023-??-??)
 
 *	Add method ``timestamp`` to UL4 type ``datetime``.
 
+*	Add attribute ``namespace`` to UL4 templates.
+
 
 Changes in 5.74 (released 2023-03-01)
 -------------------------------------