+++ docs/NEWS.rst
@@ -19,6 +19,8 @@ Changes in HEAD (released 2023-??-??)
 
 *	Add attribute ``namespace`` to UL4 templates.
 
+*	Updated UL4 and UL4ON test infrastructure to work around problems with gradle.
+
 
 Changes in 5.74 (released 2023-03-01)
 -------------------------------------